Revert "Merge pull request #24 from YOUKK/feature/Character"
This reverts commit 6138ec4aea5c6ab77ab80c5c61429cf15a87e9de, reversing
changes made to 00e78191866303fe3ee1331854f73545c777c7cb.

+++ Assets/Scripts/Character/Projectile/Pea.cs
@@ -14,7 +14,7 @@ public class Pea : MonoBehaviour
     {
         if( collision.tag == "Enemy" )
         {
-            collision.gameObject.GetComponent<Monster>().Hit(mainCharacter.Strength);
+            collision.gameObject.GetComponent<Monster>().Hit((int)mainCharacter.Strength);
             gameObject.SetActive(false);
         }
     }
